requirements.txt
Django==5.1.5
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
geopy
folium
selenium
webdriver-manager
gunicorn
